Revert "support `TTypes['events']`"
This reverts commit 3ffa8dc219dffc01964a42d5aa8338f40cc58f33.
