+++ packages/xstate-store/test/store.test.ts
@@ -204,6 +204,38 @@ it('can be inspected', () => {
   ]);
 });
 
+it('emits (types)', () =>
+  new Promise<void>((res) => {
+    const store = createStore({
+      types: {
+        emitted: {} as
+          | { type: 'increased'; upBy: number }
+          | { type: 'decreased'; downBy: number }
+      },
+      context: {
+        count: 0
+      },
+      on: {
+        inc: {
+          count: (ctx, _: {}, enq) => {
+            enq.emit({ type: 'increased', upBy: 1 });
+
+            // @ts-expect-error
+            enq.emit({ type: 'unknown' });
+            return ctx.count + 1;
+          }
+        }
+      }
+    });
+
+    store.on('increased', (ev) => {
+      expect(ev).toEqual({ type: 'increased', upBy: 1 });
+      res();
+    });
+
+    store.send({ type: 'inc' });
+  }));
+
 it('emitted events can be subscribed to', () => {
   const store = createStore({
     types: {
